modulos/requirements.txt
aniso8601==9.0.0
certifi==2020.12.5
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
Flask @ file:///home/ktietz/src/ci/flask_1611932660458/work
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
itsdangerous @ file:///home/ktietz/src/ci/itsdangerous_1611932585308/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
MarkupSafe==1.1.1
mysqlclient==2.0.3
passlib==1.7.4
python-http-client==3.3.2
pytz==2021.1
PyYAML==5.4.1
sendgrid==6.6.0
six==1.15.0
starkbank-ecdsa==1.1.0
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
wincertstore==0.2
